backend/requirements.txt
fastapi>=0.95.0
uvicorn[standard]>=0.22.0
pdfminer.six>=20221105
python-docx>=0.8.11
textract>=1.6.5
spacy>=3.5.0
transformers>=4.30.0
pymupdf>=1.22.0
python-multipart>=0.0.5
openai>=1.0.0
presidio-analyzer>=2.2.34
presidio-anonymizer>=2.2.34
pydantic>=1.10.0
pandas>=1.5.0
matplotlib>=3.7.0
PyPDF2>=3.0.0
pillow>=9.5.0
cohere>=4.26
starlette-cors>=0.1.0
platformdirs>=3.5.0
wheel

# Optional, for advanced OCR/image support
pytesseract>=0.3.10
tesserocr>=2.5.2
paddleocr>=2.6.1.3
opencv-python>=4.7.0
